[v6.3.0] constants defaults 분리
